requirements.txt
python-dotenv
chromadb
sentence-transformers
